back-end/requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
awsebcli==3.19.2
black==19.10b0
boto3==1.14.28
botocore==1.19.9
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.1
dnspython==2.0.0
docutils==0.15.2
filelock==3.0.12
firebase==3.0.1
Flask==1.1.2
Flask-Cors==3.0.8
future==0.16.0
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsons==1.3.0
jwcrypto==0.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
oauth2client==4.1.3
pathspec==0.5.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.10.1
pypiwin32==223
python-dateutil==2.8.1
python-jwt==3.3.0
pywin32==228
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.3
scapy===git-archive.devd31378c886
semantic-version==2.5.0
six==1.15.0
sseclient==0.0.26
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
typish==1.9.1
urllib3==1.25.9
virtualenv==20.1.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
